requirements.txt
langchain-groq
langchain
transformers
python-dotenv